requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

boto3==1.28.62
hopsworks==3.2.0
hyperopt==0.2.7
lightgbm==4.1.0
matplotlib==3.8.0
mlflow==2.7.1
numpy==1.25.2
pandas==1.5.3
PyYAML==6.0.1
scikit-learn==1.3.1
seaborn==0.13.0
shap==0.42.1
xgboost==1.7.6
